[MIG] account_financial_discount: Migrated force feature to 18.0
