--- account_financial_discount/models/__init__.py
@@ -1,6 +1,0 @@
-from . import account_payment_term
-from . import account_move
-from . import account_reconcile_model
-from . import res_company
-from . import res_config_settings
-from . import account_move_line
