--- account_financial_discount/readme/CONFIGURE.md
@@ -1,5 +1,0 @@
-Go to Accounting \> Configuration \> Payment Terms, to define a payment
-term involving a financial discount:
-
-- Set a number of days while discount is available after posting
-- Set a discount percentage
