--- account_financial_discount/readme/DESCRIPTION.md
@@ -1,2 +1,0 @@
-This module adds functionality to handle financial discounts for early
-payments, and to generate proper write-offs during register payments.
