--- account_financial_discount/readme/ROADMAP.md
@@ -1,8 +1,0 @@
-- Storage of available discount is only done if the invoice doesn't
-  include more than one account.tax.
-- Tax write-off is not only done automatically on register payment.
-- On payment term having multiple term lines, financial discount will be
-  applied even if the statement amount covers only first payment term
-  line
-- Financial discount cannot be applied if the payment is done in another
-  currency than the invoice.
