--- account_financial_discount/readme/USAGE.md
@@ -1,7 +1,0 @@
-When a payment term with discount is selected on an invoice (i.e
-Customers invoices or Vendor bills), the available discount will be
-stored on the first payment term line once the invoice is posted.
-
-The financial discount will then be applied using register payment
-wizard if the payment is done before the discount date, or if the "Force
-financial discount" is used on the invoice.
