--- account_financial_discount/tests/__init__.py
@@ -1,4 +1,0 @@
-from . import common
-from . import test_financial_discount_manual_payment
-from . import test_financial_discount_reconciliation
-from . import test_payment_term
