--- account_financial_discount/views/account_reconcile_model.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<odoo>
-    <record id="view_account_reconcile_model_form" model="ir.ui.view">
-        <field name="name">account.reconcile.model.form</field>
-        <field name="model">account.reconcile.model</field>
-        <field name="inherit_id" ref="account.view_account_reconcile_model_form" />
-        <field name="arch" type="xml">
-            <field name="to_check" position="after">
-                <field
-                    name="apply_financial_discounts"
-                    attrs="{'invisible': [('rule_type', '!=', 'invoice_matching')]}"
-                />
-            </field>
-            <xpath expr="//sheet" position="inside">
-                <group
-                    string="Financial discount"
-                    name="financial_discount"
-                    attrs="{'invisible':['|', ('rule_type', '!=', 'invoice_matching'), ('apply_financial_discounts', '=', False)]}"
-                >
-                    <group>
-                        <field name="financial_discount_label" />
-                        <field name="financial_discount_expense_account_id" />
-                    </group>
-                    <group>
-                        <field name="financial_discount_tolerance" />
-                        <field name="financial_discount_revenue_account_id" />
-                    </group>
-                </group>
-            </xpath>
-        </field>
-    </record>
-</odoo>
