--- account_financial_discount/views/payment_receipt.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <template
-        id="report_payment_receipt_document"
-        inherit_id="account.report_payment_receipt_document"
-    >
-        <xpath expr="//table/thead/tr/th[hasclass('text-right')][2]" position="after">
-            <t
-                t-set="any_inv_not_entry"
-                t-value="any(partial[2].move_id.move_type != entry for partial in o.move_id._get_reconciled_invoices_partials())"
-            />
-            <th t-if="any_inv_not_entry" class="text-right">
-                <span>Written-off</span>
-            </th>
-        </xpath>
-        <xpath expr="//table/tbody/tr//td[hasclass('text-right')][2]" position="after">
-            <t t-set="inv_move_line_discount" t-value="rec[2].amount_discount" />
-            <td class="text-right">
-                <span
-                    t-esc="inv_move_line_discount"
-                    t-options="{'widget': 'monetary', 'display_currency': inv.currency_id}"
-                />
-            </td>
-        </xpath>
-        <xpath expr="//span[@t-esc='amount']" position="attributes">
-            <attribute name="t-esc">amount - inv_move_line_discount</attribute>
-        </xpath>
-    </template>
-</odoo>
