--- account_financial_discount/views/res_config_settings.xml
@@ -1,62 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<odoo>
-    <record id="view_account_config_settings" model="ir.ui.view">
-        <field name="name">account settings </field>
-        <field name="model">res.config.settings</field>
-        <field name="inherit_id" ref="account.res_config_settings_view_form" />
-        <field name="arch" type="xml">
-            <xpath expr="//div[@id='print_bills_payment']/../.." position="after">
-                <div
-                    class="col-12 col-lg-6 o_setting_box"
-                    groups="account.group_account_user"
-                >
-                    <div class="o_setting_left_pane" />
-                    <div class="o_setting_right_pane">
-                        <label
-                            string="Financial discount revenue account"
-                            for="financial_discount_revenue_account_id"
-                        />
-                        <span
-                            class="fa fa-lg fa-building"
-                            title="Values set here are company-specific."
-                            aria-label="Values set here are company-specific."
-                            groups="base.group_multi_company"
-                            role="img"
-                        />
-                        <div class="content-group">
-                            <div class="row mt16 ml4">
-                                <field name="financial_discount_revenue_account_id" />
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </xpath>
-            <xpath expr="//field[@name='qr_code']/../.." position="after">
-                <div
-                    class="col-12 col-lg-6 o_setting_box"
-                    groups="account.group_account_user"
-                >
-                    <div class="o_setting_left_pane" />
-                    <div class="o_setting_right_pane">
-                        <label
-                            string="Financial discount expense account"
-                            for="financial_discount_expense_account_id"
-                        />
-                        <span
-                            class="fa fa-lg fa-building-o"
-                            title="Values set here are company-specific."
-                            aria-label="Values set here are company-specific."
-                            groups="base.group_multi_company"
-                            role="img"
-                        />
-                        <div class="content-group">
-                            <div class="row mt16 ml4">
-                                <field name="financial_discount_expense_account_id" />
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </xpath>
-        </field>
-    </record>
-</odoo>
