+++ account_force_early_discount/models/__init__.py
@@ -1,0 +1,1 @@
+from . import account_move
