+++ account_force_early_discount/readme/DESCRIPTION.md
@@ -1,0 +1,2 @@
+This module allows to force early payment discounts past the validity
+date of the discount set through the payment term.
