+++ account_force_early_discount/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_financial_discount_manual_payment
