+++ account_force_early_discount/tests/common.py
@@ -1,9 +1,11 @@
 # Copyright 2020 Camptocamp SA
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl)
-from odoo.tests.common import Form, SavepointCase
+from odoo.tests import Form, TransactionCase
+
+from odoo.addons.base.tests.common import DISABLED_MAIL_CONTEXT
 
 
-class TestAccountFinancialDiscountCommon(SavepointCase):
+class TestAccountFinancialDiscountCommon(TransactionCase):
     @classmethod
     def init_invoice(
         cls,
@@ -11,7 +13,6 @@ class TestAccountFinancialDiscountCommon(SavepointCase):
         move_type,
         payment_term=None,
         invoice_date=None,
-        invoice_date_due=None,
         currency=None,
         payment_reference=None,
     ):
@@ -21,7 +22,6 @@ class TestAccountFinancialDiscountCommon(SavepointCase):
         move_form.partner_id = partner
         move_form.invoice_payment_term_id = payment_term
         move_form.invoice_date = invoice_date
-        move_form.invoice_date_due = invoice_date_due
         if currency is not None:
             move_form.currency_id = currency
         if payment_reference is not None:
@@ -39,16 +39,22 @@ class TestAccountFinancialDiscountCommon(SavepointCase):
                 line_form.name = product and product.name or "test"
                 line_form.quantity = quantity
                 line_form.price_unit = unit_price
-                if not with_tax:
-                    line_form.tax_ids.clear()
+                if with_tax:
+                    if invoice.move_type in ("out_invoice", "in_refund"):
+                        tax = cls.sale_tax
+                    elif invoice.move_type in ("in_invoice", "out_refund"):
+                        tax = cls.purchase_tax
+                    line_form.tax_ids = tax
 
     @classmethod
     def setUpClass(cls):
         super().setUpClass()
-
+        cls.env = cls.env(context=dict(cls.env.context, **DISABLED_MAIL_CONTEXT))
         cls.usd_currency = cls.env.ref("base.USD")
         cls.eur_currency = cls.env.ref("base.EUR")
+        cls.eur_currency.active = True
         cls.chf_currency = cls.env.ref("base.CHF")
+        cls.chf_currency.active = True
 
         cls.partner = cls.env["res.partner"].create(
             {"name": "Peter Muster AG", "supplier_rank": 1}
@@ -57,48 +63,30 @@ class TestAccountFinancialDiscountCommon(SavepointCase):
             {"name": "Hans Muster GmbH & Co. KG", "customer_rank": 1}
         )
 
-        cls.write_off_rev = cls.env["account.account"].create(
-            {
-                "code": "wrtrev",
-                "name": "writeoff revenue",
-                "user_type_id": cls.env.ref("account.data_account_type_expenses").id,
-                "reconcile": False,
-            }
-        )
-        cls.write_off_exp = cls.env["account.account"].create(
-            {
-                "code": "wrtexp",
-                "name": "writeoff expenses",
-                "user_type_id": cls.env.ref("account.data_account_type_expenses").id,
-                "reconcile": False,
-            }
-        )
-        cls.env.company.financial_discount_expense_account_id = cls.write_off_exp
-        cls.env.company.financial_discount_revenue_account_id = cls.write_off_rev
         cls.payment_term = cls.env["account.payment.term"].create(
             {
                 "name": "Skonto",
-                "days_discount": 10,
-                "percent_discount": 2.0,
+                "early_discount": True,
+                "discount_days": 10,
+                "discount_percentage": 2.0,
                 "line_ids": [
                     (
                         0,
                         0,
                         {
-                            "value": "balance",
-                            "days": 60,
-                            "option": "day_after_invoice_date",
+                            "value_amount": 100,
+                            "value": "percent",
+                            "nb_days": 60,
+                            "delay_type": "days_after",
                         },
                     )
                 ],
             }
         )
-        cls.payable_account = cls.env["account.account"].search(
-            [("user_type_id.name", "=", "Payable")], limit=1
-        )
-        cls.receivable_account = cls.env["account.account"].search(
-            [("user_type_id.name", "=", "Receivable")], limit=1
-        )
+
+        cls.sale_tax = cls.env.company.account_sale_tax_id
+        cls.purchase_tax = cls.env.company.account_purchase_tax_id
+
         cls.bank_journal = cls.env["account.journal"].search(
             [("company_id", "=", cls.env.company.id), ("type", "=", "bank")],
             limit=1,
@@ -112,17 +100,4 @@ class TestAccountFinancialDiscountCommon(SavepointCase):
             }
         )
 
-        cls.exp = cls.env["account.account"].create(
-            {
-                "code": "exp",
-                "name": "expenses",
-                "user_type_id": cls.env.ref("account.data_account_type_expenses").id,
-                "reconcile": True,
-            }
-        )
-
         cls.payment_thirty_net = cls.env.ref("account.account_payment_term_30days")
-
-        cls.payment_method_manual_out = cls.env.ref(
-            "account.account_payment_method_manual_out"
-        )
