+++ account_force_early_discount/views/account_move.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<odoo>
+    <record id="view_move_form_inherit" model="ir.ui.view">
+        <field name="name">account.move.form.inherit</field>
+        <field name="model">account.move</field>
+        <field name="inherit_id" ref="account.view_move_form" />
+        <field name="arch" type="xml">
+            <xpath
+                expr="//field[@name='invoice_payment_term_id']/parent::div"
+                position="after"
+            >
+                <field
+                    name="force_early_discount"
+                    invisible="not display_force_early_discount"
+                />
+            </xpath>
+        </field>
+    </record>
+</odoo>
